PutImage: Elide NoSwap for xcb_writev if it looks like a win
The previous code would always copy image data at least once before
writing it to the kernel: either into the connection's output buffer if
there is space, or into the heap if not.  In the latter case _XSend()
would trigger a flush.

You'd like to avoid the copy, and you mostly can, but there's a
tradeoff.  Copying the image data into the output buffer effectively
pipelines the operation, but writev is a synchronous call.  So we
preserve the existing heuristic so as not to trash performance for small
image writes.  (Leaving that out drops x11perf -putimage10 from
~2M/sec to ~300K/sec on one machine.)  Note that we actually pipeline if
the request would fit in a freshly allocated output buffer; that may
mean an implicit flush of the existing buffer and a new allocation, but
we'd rather keep the pipeline full.

This still sucks, because xcb_writev() has to expensively account for
bytes written, and we don't pass in the total data size to accelerate
the common case of writing everything.  (The reverse optimisation might
be worthwhile too: when the write was incomplete, the caller could more
easily determine where to resume than the callee if the size written was
returned.)

With that fixed this would probably be a rather nice win for large
image uploads.  Alternatively we could take a syscall to look ahead at
the available buffer space and only queue up enough iovec to fill it...

Signed-off-by: Adam Jackson <ajax@redhat.com>