Add French locale with compose sequences for standardised layout
The standardised AZERTY layout is defined by the French Agency for
Standardisation (AFNOR) in standard NF Z71-300 (as opposed to the legacy
AZERTY layout, which is a de-facto standard that was never formalised).

Buy document: https://www.boutique.afnor.org/en-gb/standard/nf-z71300/user-interfaces-french-keyboard-layouts-for-office/fa188960/1792
Public information: https://norme-azerty.fr/en/#explore

This layout contains a set of compositions for “other symbols used in
European Latin languages” which are added in the new Compose file with
the new dead key dead_similar.

Signed-off-by: Cimbali <me@cimba.li>