NLS: Update a dead_tilde sequence for bépo 1.0, add one from bépo 1.1
According to https://gitlab.com/bepo/pilotes/-/commits/master/xkb
the `dead_tilde` `equals` sequence for `similareq` appears to be
based on a beta version of the bépo layout, as it shows up in their
commit history prior to the formal version 1.0 release in 2008, but
was quickly changed to use `dead tilde` `minus` instead by the time
the full relase rolled out in 2009. This patch updates this sequence
to the one used in the full version since 2009. Additionally, it adds
one additional sequence from the 1.1 version of the layout from 2019
to compensate for the loss, which uses the original `dead_tilde`
`equals` sequence for the useful `approxeq` symbol.