NLS: Add sequences for characters in Enclosed Alphanumeric Supplement.
This expands upon the sequences for encircled letters and numbers
found in the original Enclosed Alphanumeric Unicode block as well as
the Enclosed CJK Letters And Months Unicode block.
Copyleft symbol is also moved to the new block for continuity.
Furthermore, An additional sequence for a special encicled letter P,
called "Sound Recording Copyright", is added in the
Letterlike Symbols Unicode block.