nls: improve compose sequences for ruble
These were originally introduced in
d9b2cc35edf01e4f751a9ccb5ac185745184cec0, and
while the logic behind them is understandable
(Cyrillic Ze in on the same key as Latin P in
QWERTY+JCUKEN layout), for someone typing in
Cyrillic, they are not really obvious: it is much
more intuitive to use Cyrillic Er to type ruble
sign (the letter it is based on). This commit
fixes that.